RUSALMA/main/filters.py
import django_filters

from .models import *


class Filter(django_filters.FilterSet):
    title = django_filters.CharFilter(field_name='title', lookup_expr='icontains')

    class Meta:
        model = Post
        fields = ['tags', 'title']

